Move reward tests into their own directory
