+++ evap/rewards/views.py
@@ -119,13 +119,14 @@ def reward_point_redemption_event_export(request, event_id):
 
 @staff_required
 def semester_activation(request, semester_id, active):
+    semester = get_object_or_404(Semester, id=semester_id)
     active = active == 'on'
 
     try:
-        activation = SemesterActivation.objects.filter(semester=Semester.objects.get(id=semester_id)).get()
+        activation = SemesterActivation.objects.filter(semester=semester).get()
         activation.is_active = active
     except SemesterActivation.DoesNotExist:
-        activation = SemesterActivation(semester=Semester.objects.get(id=semester_id), is_active=active)
+        activation = SemesterActivation(semester=semester, is_active=active)
     activation.save()
 
     return semester_view(request=request, semester_id=semester_id)
